+++ hs-bindgen/test-artefacts/fixtures/macros/macro_type_void.2.raw/Example.hs
@@ -3,11 +3,14 @@ module Example
     )
   where
 
+import qualified HsBindgen.Runtime.Macro as Macro
+import qualified HsBindgen.Runtime.Support as BG
+
 {-| __C declaration:__ @macro MyVoid@
 
     __defined at:__ @macros\/macro_type_void.h 3:9@
 
     __exported by:__ @macros\/macro_type_void.h@
 -}
-myVoid :: [String]
-myVoid = ["void"]
+myVoid :: Macro.Raw BG.Text
+myVoid = Macro.objectLike "MyVoid" ["void"]