+++ hs-bindgen/test-artefacts/fixtures/macros/parse/elaborate.2.raw/Example.hs
@@ -24,6 +24,7 @@ module Example
   where
 
 import qualified HsBindgen.Runtime.HasCField as HasCField
+import qualified HsBindgen.Runtime.Macro as Macro
 import qualified HsBindgen.Runtime.Marshal as Marshal
 import qualified HsBindgen.Runtime.Support as BG
 import qualified HsBindgen.Runtime.Support.CompatHasField as BG.CompatHasField
@@ -34,8 +35,8 @@ import qualified HsBindgen.Runtime.Support.CompatHasField as BG.CompatHasField
 
     __exported by:__ @macros\/parse\/elaborate.h@
 -}
-iNNER_A :: [String]
-iNNER_A = ["OUTER_A"]
+iNNER_A :: Macro.Raw BG.Text
+iNNER_A = Macro.objectLike "INNER_A" ["OUTER_A"]
 
 {-| __C declaration:__ @macro INNER_B@
 
@@ -43,8 +44,8 @@ iNNER_A = ["OUTER_A"]
 
     __exported by:__ @macros\/parse\/elaborate.h@
 -}
-iNNER_B :: [String]
-iNNER_B = ["2"]
+iNNER_B :: Macro.Raw BG.Text
+iNNER_B = Macro.objectLike "INNER_B" ["2"]
 
 {-| __C declaration:__ @macro OUTER_A@
 
@@ -52,8 +53,8 @@ iNNER_B = ["2"]
 
     __exported by:__ @macros\/parse\/elaborate.h@
 -}
-oUTER_A :: [String]
-oUTER_A = ["1"]
+oUTER_A :: Macro.Raw BG.Text
+oUTER_A = Macro.objectLike "OUTER_A" ["1"]
 
 {-| __C declaration:__ @outer_int@
 
@@ -158,8 +159,8 @@ instance HasCField.HasCField Inner_int "unwrapInner_int" where
 
     __exported by:__ @macros\/parse\/elaborate.h@
 -}
-oUTER_B :: [String]
-oUTER_B = ["INNER_A"]
+oUTER_B :: Macro.Raw BG.Text
+oUTER_B = Macro.objectLike "OUTER_B" ["INNER_A"]
 
 {-| __C declaration:__ @macro OUTER_C@
 
@@ -167,5 +168,5 @@ oUTER_B = ["INNER_A"]
 
     __exported by:__ @macros\/parse\/elaborate.h@
 -}
-oUTER_C :: [String]
-oUTER_C = ["INNER_B"]
+oUTER_C :: Macro.Raw BG.Text
+oUTER_C = Macro.objectLike "OUTER_C" ["INNER_B"]